feat: Add thumbnailContainer
