+++ src/components/shared/thumbnail/index.jsx
