+++ src/components/shared/thumbnail/styled.js
