+++ src/components/shared/thumbnailContainer/houseContainer/index.jsx
@@ -1,0 +1,42 @@
+import React from "react";
+import { useNavigate } from "react-router-dom";
+import { StImgWrap, StThumbnailContainer } from "../../../pages/home/styled";
+import Thumbnail from "../../thumbnail";
+import NumberComma from "../../numberComma";
+import { home_img_01 } from "../../../../asset/images/sample";
+
+function HouseContainer({ articleList, column }) {
+  const navigate = useNavigate();
+  const onClickArticleHandler = (articleId) => {
+    navigate(`/articles/${articleId}`);
+  };
+  return (
+    <StThumbnailContainer>
+      {articleList.map((article) => (
+        <Thumbnail
+          key={article.articleId}
+          type={"house"}
+          column={column.house}
+          align={"center"}
+          onClick={() => onClickArticleHandler(article.articleId)}
+        >
+          <StImgWrap>
+            <img src={home_img_01} alt="home_img" />
+          </StImgWrap>
+          <h1>{article.title}</h1>
+          <address>{article.nickname}</address>
+          <footer>
+            <span>
+              스크랩{" "}
+              <NumberComma number="123" size={"13px"} lineheight={"17px"} /> ·
+              조회{" "}
+              <NumberComma number="456" size={"13px"} lineheight={"17px"} />
+            </span>
+          </footer>
+        </Thumbnail>
+      ))}
+    </StThumbnailContainer>
+  );
+}
+
+export default HouseContainer;
